+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Fix CCS811 Input measurement issue ([#641](https://github.com/kizniche/mycodo/issues/641))
  - Fix Math controller (equation)
  - Fix sending email notification to multiple recipients
+ - Prevent RPyC TimeoutError from crashing PID controller
 
 ### Features